TODOs done: rename, duplication
