+++ org.projectusus.ui.dependencygraph/src/org/projectusus/ui/dependencygraph/colorProvider/ClassEdgeColorProvider.java
@@ -3,25 +3,18 @@ package org.projectusus.ui.dependencygraph.colorProvider;
 import static org.projectusus.ui.colors.UsusColors.DARK_GREY;
 import static org.projectusus.ui.colors.UsusColors.DARK_RED;
 import static org.projectusus.ui.colors.UsusColors.getSharedColors;
+import static org.projectusus.ui.dependencygraph.nodes.NodeLabelProvider.isCrossPackageRelation;
 
 import org.eclipse.swt.graphics.Color;
-import org.projectusus.ui.dependencygraph.nodes.GraphNode;
+import org.projectusus.ui.dependencygraph.nodes.IEdgeColorProvider;
 
 public class ClassEdgeColorProvider implements IEdgeColorProvider {
 
     public Color getEdgeColor( Object src, Object dest, boolean hightlightStrongConnections ) {
-        if( areClassesInDifferentPackages( src, dest ) ) {
+        if( isCrossPackageRelation( src, dest ) ) {
             return getSharedColors().getColor( DARK_RED );
         }
         return getSharedColors().getColor( DARK_GREY );
-    }
-
-    // TODO aOSD kopiert von NodeLabelProvider
-    private boolean areClassesInDifferentPackages( Object src, Object dest ) {
-        if( (src instanceof GraphNode) && dest instanceof GraphNode ) {
-            return ((GraphNode)src).isInDifferentPackageThan( (GraphNode)dest );
-        }
-        return false;
     }
 
     public void refresh() {
