+++ org.projectusus.ui.dependencygraph/src/org/projectusus/ui/dependencygraph/common/DependencyGraphViewer.java
@@ -27,8 +27,10 @@ import org.eclipse.zest.core.viewers.GraphViewer;
 import org.eclipse.zest.core.widgets.Graph;
 import org.eclipse.zest.core.widgets.GraphItem;
 import org.eclipse.zest.core.widgets.ZestStyles;
-import org.projectusus.ui.dependencygraph.colorProvider.IEdgeColorProvider;
 import org.projectusus.ui.dependencygraph.nodes.GraphNode;
+import org.projectusus.ui.dependencygraph.nodes.IEdgeColorProvider;
+import org.projectusus.ui.dependencygraph.nodes.NodeContentProvider;
+import org.projectusus.ui.dependencygraph.nodes.NodeLabelProvider;
 import org.projectusus.ui.util.EditorOpener;
 
 public class DependencyGraphViewer extends GraphViewer {
